cellController- Read, Delete 구현

+++ lab/capstone-cell-service/src/main/java/com/example/capstonecellservice/dto/CellDto.java
@@ -10,7 +10,7 @@ public class CellDto {
     private Long totalCell;
     private Long liveCell;
     private Long deadCell;
-    private Long viability;
+    private Double viability;
 
     private String userId;
 }
